+++ page.php
@@ -11,6 +11,13 @@
  * @since CNRS Web Kit 1.0
  */
 
+global $cnrs_global_params;
+$sidebar = $cnrs_global_params->field('page_with_sidebar');
+
+if (! $sidebar){
+    add_filter( 'body_class', 'add_no_sidebar_class' );
+}
+
 get_header(); ?>
 
 <div id="primary" class="content-area">
@@ -35,5 +42,9 @@ get_header(); ?>
 	<?php get_sidebar( 'content-bottom' ); ?>
 
 </div><!-- .content-area -->
-<?php get_sidebar(); ?>
-<?php get_footer(); ?>
+<?php 
+if ( $sidebar ){
+    get_sidebar();
+}
+
+get_footer();