+++ single-evenement.php
@@ -7,6 +7,13 @@
  * @since CNRS Web Kit 1.0
  */
 
+global $cnrs_global_params;
+$sidebar = $cnrs_global_params->field('evenement_with_sidebar');
+
+if (! $sidebar){
+    add_filter( 'body_class', 'add_no_sidebar_class' );
+}
+
 get_header(); ?>
 
 <div id="primary" class="content-area">
@@ -49,6 +56,9 @@ get_header(); ?>
 	<?php get_sidebar( 'content-bottom' ); ?>
 
 </div><!-- .content-area -->
+<?php 
+if ( $sidebar ){
+    get_sidebar();
+}
 
-<?php get_sidebar(); ?>
-<?php get_footer(); ?>
+get_footer();